requirements.txt
keras==2.3.1
scikit-learn==0.21.3
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
pyyaml==5.2.0
kaleido==0.2.1
tensorflow==2.0.0
faiss-cpu==1.7.0
py-cpuinfo==8.0.0
psutil==5.8.0
